Commit 7 - Smooth working camera and tweened inventory
Inventory UI has been added and is working in scene with a tween. Uses a tween manager to cast its setup method to the player connector which is triggered by an input.
Camera had a bug where is bonced and wasn't looking at centre mass of character on lock. This has now been fixed and I think the camera is working smoothly for now.